requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Async support
asyncio-throttle>=1.0.2

# Utility libraries
python-multipart>=0.0.6
python-json-logger>=2.0.7

# Optional: For enhanced logging and monitoring
prometheus-client>=0.19.0
structlog>=23.2.0
